requirements.txt
python-telegram-bot[persistence]
numexpr
requests
python-dotenv